login,signup,editprofile 컴포넌트 리펙토링
