+++ src/components/BlueBoard.tsx
@@ -12,7 +12,7 @@ export default function BlueBoard({
     <>
       <div
         className={twMerge(
-          "w-full max-w-[500px] min-w-[288px] rounded-[5px] p-[35px] shadow-[0_0_4px_rgba(0,51,160,0.25)] mb-[100px]",
+          "w-full max-w-[500px] min-w-[288px] rounded-[5px] p-[35px] shadow-[0_0_4px_rgba(0,51,160,0.25)]",
           className
         )}
       >
